break out some functions, new list routes
