+++ serve/src/main.rs
@@ -1,131 +1,20 @@
 use actix_web::web::{Data, Query};
 use actix_web::{get, middleware, App, HttpResponse, HttpServer};
 use chrono::{DateTime, Utc};
+use db_util::{get_all_platforms, get_platform_by_name, market, platform, Market};
 use diesel::r2d2::{ConnectionManager, Pool, PooledConnection};
 use diesel::{pg::PgConnection, prelude::*};
+use helper::{scale_list, ApiError};
+use market_calibration::{build_calibration_plot, CalibrationQueryParams};
+use market_filter::{get_markets_filtered, CommonFilterParams};
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use std::env::var;
 
+mod db_util;
 mod helper;
-use helper::ApiError;
-
-// Diesel macro to get markets from the database table.
-table! {
-    market (id) {
-        id -> Int4,
-        title -> Varchar,
-        platform -> Varchar,
-        platform_id -> Varchar,
-        url -> Varchar,
-        open_dt -> Timestamptz,
-        close_dt -> Timestamptz,
-        open_days -> Float,
-        volume_usd -> Float,
-        num_traders -> Integer,
-        category -> Varchar,
-        prob_at_midpoint -> Float,
-        prob_at_close -> Float,
-        prob_time_weighted -> Float,
-        resolution -> Float,
-    }
-}
-table! {
-    platform (platform_name) {
-        platform_name -> Varchar,
-        platform_name_fmt -> Varchar,
-        platform_description -> Varchar,
-        platform_avatar_url -> Varchar,
-        platform_color -> Varchar,
-    }
-}
-
-/// Data returned from the database, same as what we inserted.
-#[derive(Debug, Queryable, Serialize, Selectable)]
-#[diesel(table_name = market)]
-struct Market {
-    title: String,
-    platform: String,
-    platform_id: String,
-    url: String,
-    open_dt: DateTime<Utc>,
-    close_dt: DateTime<Utc>,
-    open_days: f32,
-    volume_usd: f32,
-    num_traders: i32,
-    category: String,
-    prob_at_midpoint: f32,
-    prob_at_close: f32,
-    prob_time_weighted: f32,
-    resolution: f32,
-}
-
-/// Data about a platform cached in the database.
-#[derive(Debug, Queryable, Serialize, Selectable)]
-#[diesel(table_name = platform)]
-struct Platform {
-    platform_name: String,
-    platform_name_fmt: String,
-    platform_description: String,
-    platform_avatar_url: String,
-    platform_color: String,
-}
-
-/// Filter parameters common to all queries.
-/// Numeric filters are formatted as Vecs:
-/// - If the list is empty or not supplied, no filtering is done.
-/// - If the list has one parameter, it is taken as the minimum.
-/// - If the list has more than one parameter, the last is taken as the maximum.
-/// - All numeric values in the schema are positive, so if you want to only limit
-///   by the maximum, set the first value to -1.
-#[derive(Debug, Deserialize, Clone)]
-struct CommonFilterParams {
-    title_contains: Option<String>,
-    platform_select: Option<String>,
-    category_select: Option<String>,
-    #[serde(default)]
-    open_dt: Vec<DateTime<Utc>>,
-    #[serde(default)]
-    close_dt: Vec<DateTime<Utc>>,
-    #[serde(default)]
-    open_days: Vec<f32>,
-    #[serde(default)]
-    volume_usd: Vec<f32>,
-    #[serde(default)]
-    num_traders: Vec<i32>,
-    #[serde(default)]
-    prob_at_midpoint: Vec<f32>,
-    #[serde(default)]
-    prob_at_close: Vec<f32>,
-    #[serde(default)]
-    prob_time_weighted: Vec<f32>,
-    #[serde(default)]
-    resolution: Vec<f32>,
-}
-
-/// Parameters passed to the calibration function.
-/// If the parameter is not supplied, the default values are used.
-/// TODO: Change calibration_bin_method and calibration_weight_attribute to enums
-#[derive(Debug, Deserialize)]
-struct CalibrationQueryParams {
-    #[serde(default = "default_calibration_bin_method")]
-    calibration_bin_method: String,
-    #[serde(default = "default_calibration_bin_size")]
-    calibration_bin_size: f32,
-    #[serde(default = "default_calibration_weight_attribute")]
-    calibration_weight_attribute: String,
-    #[serde(flatten)]
-    filters: CommonFilterParams,
-}
-fn default_calibration_bin_method() -> String {
-    "prob_time_weighted".to_string()
-}
-fn default_calibration_bin_size() -> f32 {
-    0.05
-}
-fn default_calibration_weight_attribute() -> String {
-    "none".to_string()
-}
+mod market_calibration;
+mod market_filter;
 
 /// Metadata to help label a plot.
 #[derive(Debug, Serialize)]
@@ -134,6 +23,7 @@ struct PlotMetadata {
     x_title: String,
     y_title: String,
 }
+
 /// Data sent to the client to render a plot, one plot per platform.
 #[derive(Debug, Serialize)]
 struct Trace {
@@ -151,158 +41,42 @@ struct Trace {
 
 /// Full response for a calibration plot.
 #[derive(Debug, Serialize)]
-struct CalibrationPlot {
+struct PlotData {
     metadata: PlotMetadata,
     traces: Vec<Trace>,
 }
 
-/// A quick and dirty f32 mask into u32 for key lookup.
-/// Only needs to work for values between 0 and 1 at increaments of 0.01.
-fn prob_to_k(f: &f32) -> i32 {
-    (f * 1000.0) as i32
+#[get("/list_platforms")]
+async fn list_platforms(
+    pool: Data<Pool<ConnectionManager<PgConnection>>>,
+) -> Result<HttpResponse, ApiError> {
+    // get database connection from pool
+    let conn = &mut pool
+        .get()
+        .map_err(|e| ApiError::new(500, format!("failed to get connection from pool: {e}")))?;
+
+    // get all platforms from database
+    let platforms = get_all_platforms(conn)?;
+
+    // send to client
+    Ok(HttpResponse::Ok().json(platforms))
 }
 
-/// Inverse of the above function.
-fn k_to_prob(k: &i32) -> f32 {
-    *k as f32 / 1000.0
-}
+#[get("/list_markets")]
+async fn list_markets(
+    query: Query<CommonFilterParams>,
+    pool: Data<Pool<ConnectionManager<PgConnection>>>,
+) -> Result<HttpResponse, ApiError> {
+    // get database connection from pool
+    let conn = &mut pool
+        .get()
+        .map_err(|e| ApiError::new(500, format!("failed to get connection from pool: {e}")))?;
 
-/// Get information about a platform from the database to send to the client.
-fn get_platform_info(
-    conn: &mut PooledConnection<ConnectionManager<PgConnection>>,
-    platform_req: &String,
-) -> Result<Platform, ApiError> {
-    use crate::platform::dsl::*;
-    let query = platform.find(&platform_req).first(conn);
-    match query {
-        Ok(p) => Ok(p),
-        Err(_) => Err(ApiError::new(
-            500,
-            format!("could not find platform data for {platform_req}"),
-        )),
-    }
-}
+    // get markets from database
+    let markets = get_markets_filtered(conn, query.into_inner())?;
 
-/// Scale a list of weights down to valid point sizes.
-fn scale_list(list: Vec<f32>, output_min: f32, output_max: f32, output_default: f32) -> Vec<f32> {
-    if list.is_empty() {
-        return Vec::new();
-    }
-
-    let input_min = *list
-        .iter()
-        .min_by(|a, b| a.partial_cmp(b).unwrap())
-        .unwrap();
-    let input_max = *list
-        .iter()
-        .max_by(|a, b| a.partial_cmp(b).unwrap())
-        .unwrap();
-
-    if input_min == input_max {
-        return vec![output_default; list.len()];
-    }
-
-    list.iter()
-        .map(|&value| {
-            ((value - input_min) / (input_max - input_min)) * (output_max - output_min) + output_min
-        })
-        .collect()
-}
-
-fn get_markets_filtered(
-    conn: &mut PooledConnection<ConnectionManager<PgConnection>>,
-    params: CommonFilterParams,
-) -> Result<Vec<Market>, ApiError> {
-    let mut query = market::table.into_boxed();
-
-    if let Some(title_contains) = params.title_contains {
-        query = query.filter(market::title.ilike(title_contains))
-    }
-
-    if let Some(platform_select) = params.platform_select {
-        query = query.filter(market::platform.eq(platform_select))
-    }
-
-    if let Some(category_select) = params.category_select {
-        query = query.filter(market::category.eq(category_select))
-    }
-
-    if let Some(min) = params.open_dt.first() {
-        query = query.filter(market::open_dt.ge(min))
-    }
-    if params.open_dt.len() > 1 {
-        let max = params.open_dt.last().unwrap();
-        query = query.filter(market::open_dt.le(max))
-    }
-
-    if let Some(min) = params.close_dt.first() {
-        query = query.filter(market::close_dt.ge(min))
-    }
-    if params.close_dt.len() > 1 {
-        let max = params.close_dt.last().unwrap();
-        query = query.filter(market::close_dt.le(max))
-    }
-
-    if let Some(min) = params.open_days.first() {
-        query = query.filter(market::open_days.ge(min))
-    }
-    if params.open_days.len() > 1 {
-        let max = params.open_days.last().unwrap();
-        query = query.filter(market::open_days.le(max))
-    }
-
-    if let Some(min) = params.volume_usd.first() {
-        query = query.filter(market::volume_usd.ge(min))
-    }
-    if params.volume_usd.len() > 1 {
-        let max = params.volume_usd.last().unwrap();
-        query = query.filter(market::volume_usd.le(max))
-    }
-
-    if let Some(min) = params.num_traders.first() {
-        query = query.filter(market::num_traders.ge(min))
-    }
-    if params.num_traders.len() > 1 {
-        let max = params.num_traders.last().unwrap();
-        query = query.filter(market::num_traders.le(max))
-    }
-
-    if let Some(min) = params.prob_at_midpoint.first() {
-        query = query.filter(market::prob_at_midpoint.ge(min))
-    }
-    if params.prob_at_midpoint.len() > 1 {
-        let max = params.prob_at_midpoint.last().unwrap();
-        query = query.filter(market::prob_at_midpoint.le(max))
-    }
-
-    if let Some(min) = params.prob_at_close.first() {
-        query = query.filter(market::prob_at_close.ge(min))
-    }
-    if params.prob_at_close.len() > 1 {
-        let max = params.prob_at_close.last().unwrap();
-        query = query.filter(market::prob_at_close.le(max))
-    }
-
-    if let Some(min) = params.prob_time_weighted.first() {
-        query = query.filter(market::prob_time_weighted.ge(min))
-    }
-    if params.prob_time_weighted.len() > 1 {
-        let max = params.prob_time_weighted.last().unwrap();
-        query = query.filter(market::prob_time_weighted.le(max))
-    }
-
-    if let Some(min) = params.resolution.first() {
-        query = query.filter(market::resolution.ge(min))
-    }
-    if params.resolution.len() > 1 {
-        let max = params.resolution.last().unwrap();
-        query = query.filter(market::resolution.le(max))
-    }
-
-    query
-        .select(Market::as_select())
-        .load::<Market>(conn)
-        .map_err(|e| ApiError::new(500, format!("failed to query markets: {e}")))
+    // send to client
+    Ok(HttpResponse::Ok().json(markets))
 }
 
 #[get("/calibration_plot")]
@@ -315,168 +89,11 @@ async fn calibration_plot(
         .get()
         .map_err(|e| ApiError::new(500, format!("failed to get connection from pool: {e}")))?;
 
-    // get all markets from database
+    // get markets from database
     let markets = get_markets_filtered(conn, query.filters.clone())?;
 
-    // sort all markets based on the platform
-    // this is a hot loop since we iterate over all markets
-    let mut markets_by_platform: HashMap<String, Vec<Market>> = HashMap::new();
-    for market in markets {
-        if let Some(market_list) = markets_by_platform.get_mut(&market.platform) {
-            market_list.push(market);
-        } else {
-            markets_by_platform.insert(market.platform.clone(), Vec::from([market]));
-        }
-    }
-
-    // generate the x-value bins
-    // note that we use u32 here instead of f32 since floating points are hard to use as keys
-    let bin_size: i32 = prob_to_k(&query.calibration_bin_size);
-    let bin_look = bin_size / 2;
-    let mut bins: Vec<i32> = Vec::new();
-    let mut x = bin_look;
-    while x <= prob_to_k(&1.0) {
-        bins.push(x);
-        x += bin_size;
-    }
-
-    let mut traces = Vec::new();
-    for (platform, market_list) in markets_by_platform {
-        // build sums and counts to use as rolling averages
-        let mut weighted_resolution_sums = HashMap::with_capacity(bins.len());
-        let mut weighted_counts = HashMap::with_capacity(bins.len());
-        // populate each map with x-values from bins
-        for bin in bins.clone() {
-            let hash_value = bin;
-            weighted_resolution_sums.insert(hash_value, 0.0);
-            weighted_counts.insert(hash_value, 0.0);
-        }
-
-        // set up brier counters
-        let mut weighted_brier_sum: f32 = 0.0;
-        let mut weighted_brier_count: f32 = 0.0;
-
-        // get weighted average values for all markets
-        // this is a hot loop since we iterate over all markets
-        for market in market_list.iter() {
-            // find the closest bin based on the market's selected x value
-            let market_x_value = match query.calibration_bin_method.as_str() {
-                "prob_at_midpoint" => market.prob_at_midpoint,
-                "prob_at_close" => market.prob_at_close,
-                "prob_time_weighted" => market.prob_time_weighted,
-                _ => {
-                    return Err(ApiError::new(
-                        400,
-                        "the value provided for `bin_method` is not a valid option".to_string(),
-                    ))
-                }
-            };
-            if market_x_value.is_nan() {
-                return Err(ApiError {
-                    status_code: 500,
-                    message: format!("Market X-Value is NaN: {:?}", market),
-                });
-            };
-            let market_k = prob_to_k(&market_x_value);
-            let bin_opt = bins
-                .iter()
-                .find(|&x| x - bin_look <= market_k && market_k <= x + bin_look);
-            let bin = match bin_opt {
-                Some(bin) => bin,
-                None => {
-                    eprintln!(
-                        "failed to find correct bin for {} in {:?} with lookaround {}",
-                        market_k, bins, bin_look
-                    );
-                    continue;
-                }
-            };
-
-            // get the weighting value
-            let market_y_value = market.resolution;
-            if market_y_value.is_nan() {
-                return Err(ApiError {
-                    status_code: 500,
-                    message: format!("Market Y-Value is NaN: {:?}", market),
-                });
-            };
-            let weight: f32 = match query.calibration_weight_attribute.as_str() {
-                "open_days" => market.open_days,
-                "num_traders" => market.num_traders as f32,
-                "volume_usd" => market.volume_usd,
-                "none" => 1.0,
-                _ => {
-                    return Err(ApiError::new(
-                        400,
-                        "the value provided for `weight_attribute` is not a valid option"
-                            .to_string(),
-                    ))
-                }
-            };
-
-            // add the market data to each counter
-            *weighted_resolution_sums.get_mut(bin).unwrap() += weight * market_y_value;
-            *weighted_counts.get_mut(bin).unwrap() += weight;
-            weighted_brier_sum += weight * (market_y_value - market_x_value).powf(2.0);
-            weighted_brier_count += weight;
-        }
-
-        // divide out rolling averages into a single average value
-        let x_series = bins.iter().map(k_to_prob).collect();
-        let y_series = bins
-            .iter()
-            .map(|bin| {
-                // note that NaN is serialized as None, so if `count` is 0 the point won't be shown
-                let sum = weighted_resolution_sums.get(bin).unwrap();
-                let count = weighted_counts.get(bin).unwrap();
-                sum / count
-            })
-            .collect();
-        let point_weights = bins
-            .iter()
-            .map(|bin| *weighted_counts.get(bin).unwrap())
-            .collect();
-        let point_sizes = scale_list(point_weights, 8.0, 32.0, 10.0);
-
-        // get cached platform info from database
-        let platform_info = get_platform_info(conn, &platform)?;
-
-        // save it all to the trace and push it to result
-        traces.push(Trace {
-            platform_name_fmt: platform_info.platform_name_fmt,
-            platform_description: platform_info.platform_description,
-            platform_avatar_url: platform_info.platform_avatar_url,
-            platform_color: platform_info.platform_color,
-            num_markets: market_list.len(),
-            brier_score: weighted_brier_sum / weighted_brier_count,
-            x_series,
-            y_series,
-            point_sizes,
-        })
-    }
-
-    traces.sort_unstable_by_key(|t| t.platform_name_fmt.clone());
-
-    let metadata = PlotMetadata {
-        title: format!("Calibration Plot"),
-        x_title: match query.calibration_bin_method.as_str() {
-            "prob_at_midpoint" => format!("Probability at Market Midpoint"),
-            "prob_at_close" => format!("Probability at Market Close"),
-            "prob_time_weighted" => format!("Market Time-Averaged Probability"),
-            _ => panic!("given bin_method not in x_title map"),
-        },
-        y_title: match query.calibration_weight_attribute.as_str() {
-            "open_days" => format!("Resolution, Weighted by Duration"),
-            "num_traders" => format!("Resolution, Weighted by Traders"),
-            "volume_usd" => format!("Resolution, Weighted by Volume"),
-            "none" => format!("Resolution, Unweighted"),
-            _ => panic!("given weight_attribute not in y_title map"),
-        },
-    };
-
-    let response = CalibrationPlot { metadata, traces };
-
-    Ok(HttpResponse::Ok().json(response))
+    // build the plot
+    build_calibration_plot(query, conn, markets)
 }
 
 /// Server startup tasks.
@@ -499,6 +116,8 @@ async fn main() -> Result<(), std::io::Error> {
             .app_data(Data::new(pool.clone()))
             .wrap(actix_cors::Cors::permissive())
             .wrap(middleware::Logger::default())
+            .service(list_platforms)
+            .service(list_markets)
             .service(calibration_plot)
     })
     .bind(var("HTTP_BIND").unwrap_or(String::from("0.0.0.0:7041")))?
